+++ source/wii/mgdl-image.cpp
@@ -6,8 +6,9 @@
 #include <malloc.h>
 #include <gccore.h>
 #include <zlib.h>
-// muffintrap: added <cstring> for version 2.6.1
+// muffintrap: added <cstring> and <stdlib.h> for version 0.100.0-muffintrap
 #include <cstring>
+#include <stdlib.h>
 
 #include "mgdl-wii/mgdl-config.h"
 #include "mgdl-wii/mgdl-types.h"
@@ -465,14 +466,14 @@ void gdl::Image::_PrepareImage(short xRes, short yRes) {
 
 	// Generate vertex arrays for fast image drawing
 	// (as were going to take advantage of the vertex cache)
-	if (vList == NULL) vList = memalign(32, sizeof(s16)*8);
+	if (vList == NULL) vList = aligned_alloc(32, sizeof(s16)*8);
 	((s16*)vList)[0] = 0;		((s16*)vList)[1] = 0;
 	((s16*)vList)[2] = xRes;	((s16*)vList)[3] = 0;
 	((s16*)vList)[4] = xRes;	((s16*)vList)[5] = yRes;
 	((s16*)vList)[6] = 0;		((s16*)vList)[7] = yRes;
 	DCFlushRange(vList, sizeof(s16)*8);
 
-	if (tList == NULL) tList = memalign(32, sizeof(f32)*8);
+	if (tList == NULL) tList = aligned_alloc(32, sizeof(f32)*8);
 	((f32*)tList)[0] = 0;	((f32*)tList)[1] = 0;
 	((f32*)tList)[2] = u;	((f32*)tList)[3] = 0;
 	((f32*)tList)[4] = u;	((f32*)tList)[5] = v;
@@ -741,14 +742,14 @@ bool gdl::Image::LoadImage(const char *fileName, u_int filterMode, u_int format)
 
 			// Generate vertex arrays for fast image drawing
 			// (as were going to take advantage of the vertex cache)
-			if (vList == NULL) vList = memalign(32, sizeof(s16)*8);
+			if (vList == NULL) vList = aligned_alloc(32, sizeof(s16)*8);
 			((s16*)vList)[0] = 0;					((s16*)vList)[1] = 0;
 			((s16*)vList)[2] = Texture.TXsize();	((s16*)vList)[3] = 0;
 			((s16*)vList)[4] = Texture.TXsize();	((s16*)vList)[5] = Texture.TYsize();
 			((s16*)vList)[6] = 0;					((s16*)vList)[7] = Texture.TYsize();
 			DCFlushRange(vList, sizeof(s16)*8);
 
-			if (tList == NULL) tList = memalign(32, sizeof(f32)*8);
+			if (tList == NULL) tList = aligned_alloc(32, sizeof(f32)*8);
 			((f32*)tList)[0] = 0;	((f32*)tList)[1] = 0;
 			((f32*)tList)[2] = 1;	((f32*)tList)[3] = 0;
 			((f32*)tList)[4] = 1;	((f32*)tList)[5] = 1;